big refactorisation de la publication
