+++ zds/tutorialv2/factories.py
@@ -9,7 +9,8 @@ from zds.tutorialv2.models.models_database import PublishableContent, Validation
 from zds.tutorialv2.models.models_versioned import Container, Extract
 from zds.utils.models import SubCategory, Licence
 from zds.gallery.factories import GalleryFactory, UserGalleryFactory
-from zds.tutorialv2.utils import publish_content, init_new_repo
+from zds.tutorialv2.utils import init_new_repo
+from zds.tutorialv2.publication_utils import publish_content
 
 text_content = u'Ceci est un texte bidon, **avec markown**'
